% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f01-Fastq_Processing.R
\name{auto_split}
\alias{auto_split}
\title{Divide FASTQ Files According to RAM Constraints}
\usage{
auto_split(
  in_fastq,
  size = 284289,
  prefix = "split_",
  compression_level = -1,
  threads = 1,
  out_dir = paste0(dirname(in_fastq), "/split"),
  extension = NULL,
  target_ram_size_gb = 10
)
}
\arguments{
\item{in_fastq}{character path to input fastq file.}

\item{size}{numeric number of reads to split larger file into
smaller files.}

\item{prefix}{character prefix for output files 
(file label, default "split_").}

\item{compression_level}{numeric compression level for
output files.}

\item{threads}{numeric number of threads to use.}

\item{out_dir}{character path to output directory.}

\item{extension}{character extension for output files (fastq | fastq.gz).}

\item{target_ram_size_gb}{numeric target size of ram in gigabytes
which each loaded file will occupy (threads * target_ram_size_gb) < ram.}
}
\value{
full path list of files generated by auto_split
}
\description{
Automatically split fastq file into smaller files by number of reads
in order to allow R instances to process the data without
segmentation faults
}
\details{
The 'auto_split' function automatically splits a FASTQ file into smaller 
files based on the available RAM size. It first calculates the number of files needed 
and the number of reads per file to ensure that each file's size fits within the 
specified 'target_ram_size_gb'. It uses the 'fastq_stats' function to obtain statistics 
on the total number of bases and reads in the input FASTQ file and estimates the 
number of files required for the target RAM size. If more than one file is needed, 
it calls the 'split_fastq' function to perform the actual splitting, applying the 
specified 'size', 'prefix', 'compression_level', and 'threads' for the operation. The function 
finally returns a list containing the file statistics and the paths to the created 
split files.
}
