% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fG-helper_functions.R
\name{combine_fasta}
\alias{combine_fasta}
\title{Combine FASTA files}
\usage{
combine_fasta(
  list_of_fasta = NULL,
  write_files = TRUE,
  out_dir = file.path(getwd(), "combined_reads"),
  prefix = "combined_telo",
  return_reads = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{list_of_fasta}{A character vector specifying the paths to the FASTA files.}

\item{write_files}{A logical value specifying whether to write the combined FASTA
file to disk. Default is TRUE.}

\item{out_dir}{A character string specifying the output directory. Default is
"combined_reads".}

\item{prefix}{A character string specifying the prefix for the output file. Default
is "combined_telo".}

\item{return_reads}{A logical value specifying whether to return the combined reads.
Default is FALSE.}

\item{verbose}{A logical value specifying whether to display verbose messages. Default
is TRUE.}
}
\value{
A list containing the combined reads.
}
\description{
Combine multiple FASTA files into a single FASTA file.
}
