% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fG-helper_functions.R
\name{estimate_ram_usage}
\alias{estimate_ram_usage}
\title{Estimate RAM usage}
\usage{
estimate_ram_usage(files, verbose = FALSE)
}
\arguments{
\item{files}{A character vector specifying the paths to the files.}

\item{verbose}{A logical value specifying whether to display verbose messages.
Default is FALSE.}
}
\value{
A list containing the mean RAM usage per file, a data frame of RAM usage
per file, and the suggested number of files to process.
}
\description{
Estimate the RAM usage of reading in a list of files.
}
