% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f01-Fastq_Processing.R
\name{fastq_stats}
\alias{fastq_stats}
\title{Generate FASTQ File Statistics and Splitting Plan}
\usage{
fastq_stats(in_fastq, threads = 1, target_ram_size_gb = 10)
}
\arguments{
\item{in_fastq}{Path to the input FASTQ file.}

\item{threads}{Number of threads to use for processing (default is 1).}

\item{target_ram_size_gb}{Desired RAM size in GB for processing (default is 10 GB).}
}
\value{
Returns a list of base stats, estimated files, and read counts.
}
\description{
Generate stats about intial fastq file and determine
how to split the file into smaller files based on RAM constraints
}
\details{
The 'fastq_stats' function analyzes a FASTQ file 
to generate basic statistics and estimates a file-splitting 
strategy based on target RAM size. It begins by constructing 
and executing a command-line call to the seqkit stats tool, 
which collects statistics such as the total number of sequences 
(reads) and the total length of sequences (bases). The function 
parses this information to extract the number of reads and 
the sum of base pairs, then calculates the total gigabases of 
data. Using the target RAM size and a predefined ratio for RAM 
usage efficiency, the function estimates the optimal number of 
smaller FASTQ files required to stay within the specified RAM 
limit and determines how many reads each of these files should 
contain. Finally, the function returns a list with statistics on 
the total number of bases, reads, the estimated number of files, 
the number of reads per file, and the total gigabases of data, 
which helps in planning the processing of large FASTQ files.
}
