% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fG-helper_functions.R
\name{find_sizes}
\alias{find_sizes}
\title{Find object sizes}
\usage{
find_sizes(clean = FALSE, which = "all", environ = .GlobalEnv)
}
\arguments{
\item{clean}{A logical value specifying whether to clean the environment. Default
is FALSE.}

\item{which}{A character string specifying which objects to remove. Default is "all".}

\item{environ}{An environment object specifying the environment to search. Default
is .GlobalEnv.
importFrom Biostrings DNAStringSet nchar substring vcountPattern}
}
\description{
Find the sizes of objects in the environment in descending order.
}
