% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f06-Telomere_length.R
\name{findTelLength}
\alias{findTelLength}
\title{findTelLength}
\usage{
findTelLength(df, st.thresh = 50, en.thresh = 50, windows_size = 5)
}
\arguments{
\item{df}{A data frame containing the telomere repeat information.}

\item{st.thresh}{A numeric value specifying the start threshold for the
telomere length calculation. Default is 50.}

\item{en.thresh}{A numeric value specifying the end threshold for the
telomere length calculation. Default is 50.}

\item{windows_size}{A numeric value specifying the size of the sliding
window. Default is 5.}
}
\value{
A data frame containing the telomere length, read length, start
and end truncation points, and the telomere end.
}
\description{
This function is used to calculate the length of the
telomere repeat in a given read.
}
