% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f05-Generate_telomere_frequencies.R
\name{generate_telomere_frequencies}
\alias{generate_telomere_frequencies}
\title{Generate telomere frequencies}
\usage{
generate_telomere_frequencies(telomere_windows, motifs)
}
\arguments{
\item{telomere_windows}{An XStringViews object containing telomere windows.}

\item{motifs}{A character vector of telomere motifs to search for.}
}
\value{
A data frame containing the start, end, window length, telomere
count, motif frequency, and hit motif.
}
\description{
Generate telomere frequencies based on each sliding window in an
individual telomere.
}
