% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fG-helper_functions.R
\name{get_mime_type}
\alias{get_mime_type}
\title{Get MIME type}
\usage{
get_mime_type(file_path)
}
\arguments{
\item{file_path}{A character string specifying the path to the file.}
}
\value{
A character string specifying the MIME type of the file.
}
\description{
Get the MIME type of a file based on its extension.
}
