% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fG-helper_functions.R
\name{get_system_ram}
\alias{get_system_ram}
\title{Get system RAM}
\usage{
get_system_ram(environment, verbose = FALSE)
}
\arguments{
\item{environment}{A character string specifying the operating system. Default
is "linux".}

\item{verbose}{A logical value specifying whether to display verbose messages.
Default is FALSE.}
}
\value{
A list containing the total amount of RAM on the system and the suggested
amount of RAM to use for processing.
}
\description{
Get the total amount of RAM on the system and suggest an amount
of RAM to use for processing.
}
