% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fshiny-launch_app.R
\name{launch_dashboard}
\alias{launch_dashboard}
\title{Launch the Package Shiny Dashboard}
\usage{
launch_dashboard()
}
\value{
A Shiny application object
}
\description{
This function launches the Shiny dashboard included in the package.
}
\examples{
if(interactive()) {
  launch_dashboard()
}
}
