% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f06-Telomere_length.R
\name{optimal_thresholds}
\alias{optimal_thresholds}
\title{Determine optimal thresholds}
\usage{
optimal_thresholds(threshold_dataframe)
}
\arguments{
\item{threshold_dataframe}{A data frame containing the start and end
thresholds for the telomere length calculation.}
}
\value{
A list containing the optimal thresholds for the telomere length
calculation.
}
\description{
Determine the optimal thresholds for the telomere length
calculation.
}
