% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f06-Telomere_length.R
\name{plot_thresholds}
\alias{plot_thresholds}
\title{plot_thresholds}
\usage{
plot_thresholds(threshold_dataframe = NULL, optimal_thresholds = NULL)
}
\arguments{
\item{threshold_dataframe}{A data frame containing the start and end
thresholds for the telomere length calculation.}

\item{optimal_thresholds}{A list containing the optimal thresholds for
the telomere length calculation.}
}
\value{
A ggplot object containing the plot of the thresholds data.
}
\description{
Plot the thresholds data with the thresholds.
}
