% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f01-Fastq_Processing.R
\name{split_fastq}
\alias{split_fastq}
\title{Divide a FASTQ File into Smaller Files}
\usage{
split_fastq(
  in_fastq,
  size = 284289,
  prefix = "split_",
  threads = 1,
  out_dir = paste0(dirname(in_fastq), "/split"),
  extension = NULL
)
}
\arguments{
\item{in_fastq}{character path to input fastq file.}

\item{size}{numeric number of reads to split larger 
fastq file into smaller files.}

\item{prefix}{character prefix for output files.}

\item{threads}{numeric number of threads to use.}

\item{out_dir}{character path to output directory.}

\item{extension}{character extension for output files.}
}
\value{
NULL, does not have a return value
}
\description{
Split fastq file into smaller files by number of reads
in order to allow R instances to process the data without
segmentation faults
}
\details{
The 'split_fastq' function is designed to divide a large 
FASTQ file into smaller files based on a specified size and includes 
options for multithreading. It constructs a 
command-line call to the 'seqkit split2' tool, which is a part of the 
SeqKit toolkit for processing sequencing data. The function takes as 
input the path to the FASTQ file to be split, the desired size 
for each split file, a prefix for naming the split files, the 
number of threads to use for parallel processing, and the directory 
where the split files will be saved. An optional file extension can 
also be specified for the split files. The function constructs the 
appropriate command with these parameters, prints the command for user 
verification, and then executes it using the system2 function to run the 
command on the system’s command line. The resulting output consists of 
multiple smaller FASTQ files, each with the specified size, which are 
saved in the designated output directory.
}
