% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fG-helper_functions.R
\docType{package}
\name{TECAT-package}
\alias{TECAT-package}
\alias{TECAT}
\title{TECAT: Telomere End Chromosome Assaying Tool}
\description{
A package for determining chromosome end specific telomere lengths using pattern 
matching with 3rd generation sequencing data.
}
\details{
The main functions of this package are:
\itemize{
  \item Telomere motif identification
  \item Pattern matching and analysis
  \item Length calculation
  \item Visualization
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jake-bioinfo/tecat}
}

}
\author{
Jake Reed and Mark Oelkuct
}
