% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f03-Pattern_searching.R
\name{telomere_matching}
\alias{telomere_matching}
\title{Search FASTQ Files for Telomere Repeats and Collect Statistics}
\usage{
telomere_matching(
  fastq_file,
  names_prefix = "telomere_",
  out_dir = getwd(),
  grep_list = c("TTAGGG", "CCCTAA"),
  return_telomeres = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{fastq_file}{A character string of the fastq file to search.}

\item{names_prefix}{A character string to prefix the names of the telomere sequences.}
}
\description{
telomere_matching takes in a fastq file and searches for telomere
repeats in the sequences. It returns a list of sequences which
contain telomere repeats and a list of statistics for each file.
}
\details{
The telomere_matching function extracts sequences containing telomere repeats 
from a FASTQ file by searching for specified motifs and computes relevant statistics. 
It reads the FASTQ file, identifies sequences with the motifs, and saves these 
sequences to a file while also providing a summary of telomere-related statistics. 
The function can return either just the statistics or both the statistics and the 
extracted sequences, depending on the return_telomeres parameter.
}
