% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f07-Truncate_Map.R
\name{truncate_file}
\alias{truncate_file}
\title{Truncate file}
\usage{
truncate_file(
  combined_telomere_file = NULL,
  results_data_frame = NULL,
  out_dir = file.path(getwd(), "truncated_reads"),
  prefix = "truncated",
  write_file = TRUE,
  return = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{combined_telomere_file}{A character string specifying the combined
telomere file.}

\item{results_data_frame}{A data frame containing the telomere length and hit
motif information.}

\item{out_dir}{A character string specifying the output directory.}

\item{prefix}{A character string specifying the prefix for the output files.}

\item{write_file}{A logical value specifying whether to write the output
files to disk. Default is TRUE.}

\item{return}{A logical value specifying whether to return the truncated
reads. Default is FALSE.}

\item{verbose}{A logical value specifying whether to display verbose
messages. Default is TRUE.}
}
\value{
A list containing the truncated reads and the truncated file list.
}
\description{
Truncate reads based on the telomere length and hit motif.
}
