% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f07-Truncate_Map.R
\name{truncation}
\alias{truncation}
\title{Truncate reads}
\usage{
truncation(reads, result_df)
}
\arguments{
\item{reads}{A DNAStringSet object containing reads to be truncated}

\item{result_df}{A data frame containing the telomere length and hit motif
information.}
}
\value{
A DNAStringSet object containing truncated reads.
}
\description{
Truncate reads based on the telomere length and hit motif.
}
