% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f05-Generate_telomere_frequencies.R
\name{frequencies}
\alias{frequencies}
\title{Generate telomere frequencies}
\usage{
frequencies(
  windows,
  motifs,
  environment = "linux",
  parallel = FALSE,
  threads = 0,
  save_files = TRUE,
  out_dir = file.path(getwd(), "telomere_frequencies"),
  progress = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{motifs}{A character vector of telomere motifs to search for.}

\item{telomere_windows}{An XStringViews object containing telomere windows.}
}
\value{
A data frame containing the start, end, window length, telomere
count, motif frequency, and hit motif.
}
\description{
Generate telomere frequencies based on each sliding window in an
individual telomere.
}
