% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f07-Truncate_Map.R
\name{map}
\alias{map}
\title{Map reads to reference}
\usage{
map(
  fasta = NULL,
  results_data_frame = NULL,
  reference_file = NULL,
  preset_string = "map-hifi",
  out_dir = file.path(getwd(), "mapped_reads"),
  prefix = "mapped",
  threads = 1,
  return_mapped = TRUE,
  underscore = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{fasta}{A character string specifying the path to the fasta file.}

\item{results_data_frame}{A data frame containing the telomere length and hit
motif information.}

\item{reference_file}{A character string specifying the path to the reference
genome.}

\item{preset_string}{A character string specifying the preset to use for
minimap2. Default is "map-hifi".}

\item{out_dir}{A character string specifying the output directory.}

\item{prefix}{A character string specifying the prefix for the output files.}

\item{threads}{An integer specifying the number of parallel processors to use.
Default is 1.}

\item{return_mapped}{A logical value specifying whether to return the mapped
reads. Default is TRUE.}

\item{underscore}{A logical value specifying whether to use underscores in
the chromosome names. Default is TRUE.}

\item{verbose}{A logical value specifying whether to display verbose messages.
Default is TRUE.}
}
\value{
A list containing the mapped reads and the results data frame.
}
\description{
Map reads to a reference genome using minimap2.
}
