% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f04-Sliding_window.R
\name{sliding_window}
\alias{sliding_window}
\title{Sliding Window Telomere Sequence Processing}
\usage{
sliding_window(telomere_sequence, window_length = 200, step = 100)
}
\arguments{
\item{window_length}{An integer specifying the length of each sliding
window. Default is 200.}

\item{step}{An integer specifying the step size for moving the window.
Default is 100.}

\item{telomere_sequences}{A DNAStringSet object containing the telomere
sequences to be processed.}
}
\value{
A list of IRanges::Views objects, each containing overlapping
windows of the original telomere sequences.
}
\description{
The 'sliding_window' function generates overlapping windows
of specified length from telomere sequences, facilitating the analysis
of smaller, manageable segments of the sequences.
}
\details{
The 'sliding_window' function processes a set of telomere
sequences by creating overlapping windows of a specified length.
For each sequence in the input list, it generates a series of windows
starting from the beginning of the sequence, moving in steps defined
by the 'step' parameter. Each window has a length specified by the
'window_length' parameter. The windows are stored in an
'IRanges::Views' object, with each window named according to the
original sequence. After processing all sequences, the function
cleans up memory and returns the list of windowed views, allowing
for detailed analysis of these smaller sequence segments.
}
