% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f07-Truncate_Map.R
\name{tecat_plot}
\alias{tecat_plot}
\title{Plotted mapped reads}
\usage{
tecat_plot(
  mapped_output = NULL,
  out_dir = file.path(getwd(), "mapped_reads"),
  prefix = "mapped",
  save_plots = TRUE,
  return = FALSE
)
}
\arguments{
\item{out_dir}{A character string specifying the output directory.}

\item{prefix}{A character string specifying the prefix for the output files.}

\item{save_plots}{A logical value specifying whether to save the plot to disk.}

\item{return}{A logical value specifying whether to return the plotted reads.
Default is FALSE.}

\item{mapped_results}{A list containing the mapped reads and the results data frame.}

\item{verbose}{A logical value specifying whether to display verbose messages.}
}
\value{
A list containing the histogram and violin plot.
}
\description{
Plot mapped reads.
}
