% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f02-Reference_Processing.R
\name{telomere_motif}
\alias{telomere_motif}
\title{Determine telomere motifs from reference fasta file}
\usage{
telomere_motif(
  data_ref,
  reference_telo_length = 2000,
  telo_motif_length = 6,
  number_of_motifs = 5,
  number_of_repeats = 3,
  threads = 1,
  platform = "PB",
  seed = 9273
)
}
\arguments{
\item{data_ref}{DNAStringSet reference data.}

\item{reference_telo_length}{numeric predicted length of telomeres
present in the reference file.}

\item{telo_motif_length}{numeric length of telomere motif.}

\item{number_of_motifs}{numeric number of motifs to identify.}

\item{number_of_repeats}{numeric number of repeats to identify.}

\item{threads}{numeric number of threads MEME will use.}

\item{platform}{character sequencing platform used (PB | ONT).}

\item{seed}{numeric seed for random number generation.}
}
\value{
list object which contains the list elements grep_list and
motifs
}
\description{
Identifies telomere motifs from a reference FASTA file and generates patterns for matching in FASTQ files.
}
\details{
The 'telomere_motif' function identifies telomere motifs from a reference DNA sequence dataset and 
generates search patterns for FASTQ files. It first validates the input data and extracts the telomere 
regions from the start and end of each sequence. The function then uses the MEME tool to discover telomere 
motifs and processes these motifs into various forms (complement, reverse, etc.), adding known ONT 
basecalling errors if applicable. It repeats and extends the motifs to create search patterns for 
FASTQ files. The function returns a list with the search patterns ('grep_list'), identified motifs 
('motifs'), and MEME analysis results ('results').
}
